refactor(frontend): split landing/table.rs into table/ submodule (#574)
`frontend/src/pages/landing/table.rs` was sitting at 799 lines (right at
the ~800-line file soft cap from `.claude/rules/05-rust-style.md`) with
13 `#[component]` functions plus 4 helpers packed in. Convert it to a
`table/` submodule split by responsibility — no logic or prop signature
changes.

## Files

- `frontend/src/pages/landing/table/mod.rs` (148 lines) — `EditField`,
  `BookTableContext`, `BookTable`, `SortableHeader`.
- `frontend/src/pages/landing/table/row.rs` (290 lines) — `EbookRow`,
  `use_row_state`, `RowDisplay`, `derive_row_display`,
  `EbookRowMarkup`, `EbookRowCells`.
- `frontend/src/pages/landing/table/cells.rs` (396 lines) —
  `RowTitleCell`, `RowSeriesCell`, `RowScalarCell`,
  `EbookRowCoverCell`, `EbookRowFormatsCell`, `EditableCell`,
  `AuthorsCell`, plus the `build_save_field` / `build_save_authors`
  callback factories.

`landing/mod.rs` already declared `mod table;` so the rename to
`table/mod.rs` is path-compatible — `super::table::{BookTable,
BookTableContext}` continues to resolve from
`landing/sections.rs`.

## Visibility

- `BookTable` and `BookTableContext` remain `pub(super)` (visible to
  `landing`) — no API surface change.
- Within `table/`, items the siblings exchange (`EbookRow`,
  `RowTitleCell`, `build_save_field`, ...) are `pub(super)` — visible
  inside `table` only.
- `EditField` and `RowDisplay` stay private — sibling modules see them
  through Rust's "private items visible to descendant modules" rule.

## Verification

- `cargo fmt --check` — clean
- `cargo clippy -p omnibus-frontend --features server --all-targets --
  -D warnings` — clean
- `cargo test -p omnibus-frontend --features server` — 172 passed, 0
  failed (unchanged from main)
- No new `#[allow(...)]` suppressions.

Closes #574
