--- frontend/src/pages/landing/table.rs
@@ -1,799 +1,0 @@
-//! Power-user table view for the landing page.
-//!
-//! Dense `<table>` rendering with inline-editable authors / title / tags
-//! for admins. Used by [`super::LandingPage`] when the view-mode toggle is
-//! set to table.
-
-use dioxus::prelude::*;
-use dioxus_router::use_navigator;
-use omnibus_shared::{Contributor, EbookMetadata, MetadataOverrides, SortDir, SortKey, ViewPrefs};
-
-use super::filtering::format_badge_label;
-use super::sorting::{contributor_names, row_slug};
-use crate::components::chip_editor::{ChipEditor, SuggestionItem};
-use crate::{data, Route};
-
-/// Inline-editable field on a power-user-table row. Used by `EbookRow` to
-/// track which of its cells (if any) is in edit mode at any given time —
-/// single-cell-at-a-time keeps the keyboard/blur lifecycle simple and
-/// avoids fighting the row's click-to-navigate handler.
-///
-/// `Series` edits the series *name* only — the series index ("#N") stays
-/// in the full metadata edit page on `/books/:uuid/edit` to avoid adding a
-/// separate Series Index column to the power-user table. Title + Series +
-/// Authors covers the stated cleanup pain (stripping series-prefix cruft
-/// from titles, renaming author/series variants).
-#[derive(Clone, Copy, PartialEq, Eq, Debug)]
-enum EditField {
-    Title,
-    Series,
-    Publisher,
-    Published,
-    Language,
-    Authors,
-}
-
-/// Per-page context threaded through `BookTable` → `EbookRow` → cells.
-#[derive(Clone, PartialEq)]
-pub(super) struct BookTableContext {
-    pub(super) server_url: String,
-    pub(super) is_admin: bool,
-    pub(super) author_suggestions: ReadSignal<Vec<SuggestionItem>>,
-    pub(super) tag_suggestions: ReadSignal<Vec<SuggestionItem>>,
-}
-
-#[component]
-pub(super) fn BookTable(
-    books: Vec<EbookMetadata>,
-    prefs: ViewPrefs,
-    on_sort: EventHandler<SortKey>,
-    ctx: BookTableContext,
-) -> Element {
-    rsx! {
-        div {
-            id: "ebook-table",
-            "data-testid": "ebook-table",
-            class: "ebook-table-wrap",
-            table { class: "ebook-table",
-                thead {
-                    tr {
-                        th { class: "ebook-col-cover", "Cover" }
-                        SortableHeader {
-                            class: "ebook-col-title".to_string(),
-                            label: "Title".to_string(),
-                            sort_key: SortKey::Title,
-                            prefs: prefs.clone(),
-                            on_sort: on_sort,
-                        }
-                        SortableHeader {
-                            class: "ebook-col-author".to_string(),
-                            label: "Author".to_string(),
-                            sort_key: SortKey::Author,
-                            prefs: prefs.clone(),
-                            on_sort: on_sort,
-                        }
-                        SortableHeader {
-                            class: "ebook-col-series".to_string(),
-                            label: "Series".to_string(),
-                            sort_key: SortKey::Series,
-                            prefs: prefs.clone(),
-                            on_sort: on_sort,
-                        }
-                        th { class: "ebook-col-publisher", "Publisher" }
-                        th { class: "ebook-col-published", "Published" }
-                        th { class: "ebook-col-formats", "Formats" }
-                        SortableHeader {
-                            class: "ebook-col-updated".to_string(),
-                            label: "Last Updated".to_string(),
-                            sort_key: SortKey::LastUpdated,
-                            prefs: prefs.clone(),
-                            on_sort: on_sort,
-                        }
-                        SortableHeader {
-                            class: "ebook-col-added".to_string(),
-                            label: "Added".to_string(),
-                            sort_key: SortKey::NewestAdded,
-                            prefs: prefs.clone(),
-                            on_sort: on_sort,
-                        }
-                        th { class: "ebook-col-language", "Language" }
-                    }
-                }
-                tbody {
-                    for book in books.into_iter() {
-                        EbookRow {
-                            key: "{book.filename}",
-                            book: book,
-                            ctx: ctx.clone(),
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-#[component]
-fn SortableHeader(
-    class: String,
-    label: String,
-    sort_key: SortKey,
-    prefs: ViewPrefs,
-    on_sort: EventHandler<SortKey>,
-) -> Element {
-    let active = prefs.sort_key == sort_key;
-    let aria_sort = match (active, prefs.sort_dir) {
-        (true, SortDir::Asc) => "ascending",
-        (true, SortDir::Desc) => "descending",
-        _ => "none",
-    };
-    let arrow = if !active {
-        ""
-    } else if prefs.sort_dir == SortDir::Asc {
-        " ↑"
-    } else {
-        " ↓"
-    };
-    rsx! {
-        th { class: "{class} sort-th", aria_sort: "{aria_sort}",
-            button {
-                class: "sort-th-btn",
-                onclick: move |_| on_sort.call(sort_key),
-                "{label}{arrow}"
-            }
-        }
-    }
-}
-
-#[component]
-fn EbookRow(book: EbookMetadata, ctx: BookTableContext) -> Element {
-    let BookTableContext {
-        server_url,
-        is_admin,
-        author_suggestions,
-        tag_suggestions,
-    } = ctx;
-    let _ = tag_suggestions; // reserved for a future inline-tag editor
-
-    let uuid = book.unique_identifier.clone().unwrap_or_default();
-    let (book_state, editing, authors_draft) = use_row_state(book.clone());
-    let save_field = build_save_field(uuid.clone(), server_url.clone(), book_state);
-    let save_authors = build_save_authors(uuid.clone(), server_url.clone(), book_state);
-    let display = derive_row_display(&book_state.read(), &server_url, &uuid);
-
-    rsx! {
-        EbookRowMarkup {
-            display,
-            uuid,
-            is_admin,
-            editing,
-            authors_draft,
-            author_suggestions,
-            save_field,
-            save_authors,
-        }
-    }
-}
-
-/// Per-row reactive state — optimistic book copy, current editing cell, and
-/// the authors chip-editor draft. The optimistic copy lets a successful
-/// inline save update the row immediately without a full library refetch;
-/// the `book` prop is resynced into it on upstream changes only when no
-/// cell is mid-edit (so a save round-trip doesn't clobber an open input).
-/// `authors_draft` mirrors `book_state.creators` so the chip editor sees
-/// canonical names after a save without dropping in-progress chip edits.
-fn use_row_state(
-    book: EbookMetadata,
-) -> (
-    Signal<EbookMetadata>,
-    Signal<Option<EditField>>,
-    Signal<Vec<String>>,
-) {
-    let mut book_state: Signal<EbookMetadata> = use_signal(|| book.clone());
-    let editing: Signal<Option<EditField>> = use_signal(|| None);
-    use_effect(use_reactive!(|book| {
-        if editing().is_none() {
-            book_state.set(book.clone());
-        }
-    }));
-
-    let initial_authors: Vec<String> = book_state
-        .read()
-        .creators
-        .iter()
-        .map(|c| c.name.clone())
-        .collect();
-    let mut authors_draft: Signal<Vec<String>> = use_signal(|| initial_authors);
-    use_effect(move || {
-        let canonical: Vec<String> = book_state
-            .read()
-            .creators
-            .iter()
-            .map(|c| c.name.clone())
-            .collect();
-        if *authors_draft.peek() != canonical {
-            authors_draft.set(canonical);
-        }
-    });
-
-    (book_state, editing, authors_draft)
-}
-
-/// Pre-derived, ready-to-render display strings for a single row. Kept in
-/// one struct so `EbookRowMarkup` reads cleanly and the row body stays
-/// under the 80-line cap.
-#[derive(Clone, PartialEq)]
-struct RowDisplay {
-    book: EbookMetadata,
-    row_testid: String,
-    thumb_base: String,
-    has_cover: bool,
-    title: String,
-    series_line: String,
-    series_text: String,
-    authors_text: String,
-    updated: String,
-    added: String,
-    publisher: String,
-    published: String,
-    language: String,
-}
-
-/// Compute the per-cell display strings from the optimistic book state.
-/// Pulled out of [`EbookRow`] so the closures and rsx live in their own
-/// scopes.
-fn derive_row_display(book: &EbookMetadata, server_url: &str, uuid: &str) -> RowDisplay {
-    let row_testid = format!("ebook-row-{}", row_slug(&book.filename));
-    let thumb_base = format!("{server_url}/api/thumbs/{uuid}");
-    let title = book.title.as_deref().unwrap_or(&book.filename).to_string();
-    let series_line = match (book.series.as_deref(), book.series_index.as_deref()) {
-        (Some(s), Some(i)) => format!("{s} #{i}"),
-        (Some(s), None) => s.to_string(),
-        _ => String::new(),
-    };
-    RowDisplay {
-        row_testid,
-        thumb_base,
-        has_cover: book.cover_url.is_some(),
-        title,
-        series_line,
-        series_text: book.series.clone().unwrap_or_default(),
-        authors_text: contributor_names(&book.creators),
-        updated: book.modified.as_deref().unwrap_or("").to_string(),
-        added: book.added_at.as_deref().unwrap_or("").to_string(),
-        publisher: book.publisher.clone().unwrap_or_default(),
-        published: book.published.clone().unwrap_or_default(),
-        language: book.language.clone().unwrap_or_default(),
-        book: book.clone(),
-    }
-}
-
-/// Build the common per-cell save callback. Empty strings clear the
-/// override (None — the scanned value re-surfaces); non-empty strings
-/// persist as `Some(value)`. The server-side merge in `rpc_save_overrides`
-/// returns the canonical merged metadata which we install verbatim into
-/// the optimistic `book_state` signal.
-fn build_save_field(
-    uuid: String,
-    server_url: String,
-    mut book_state: Signal<EbookMetadata>,
-) -> impl FnMut((EditField, String)) + 'static {
-    move |args: (EditField, String)| {
-        let (field, value) = args;
-        let uuid = uuid.clone();
-        let url = server_url.clone();
-        spawn(async move {
-            let mut overrides = MetadataOverrides::default();
-            let trimmed = value.trim();
-            let value = if trimmed.is_empty() {
-                None
-            } else {
-                Some(trimmed.to_string())
-            };
-            match field {
-                EditField::Title => overrides.title = value,
-                EditField::Series => overrides.series = value,
-                EditField::Publisher => overrides.publisher = value,
-                EditField::Published => overrides.published = value,
-                EditField::Language => overrides.language = value,
-                // Authors edits go through `build_save_authors` so they
-                // can set the `creators` override instead of a scalar.
-                EditField::Authors => return,
-            }
-            if overrides.validate().is_err() {
-                // F5.1 length caps rejected the payload. The display
-                // reverts on the next refetch; a toast / inline error
-                // message ships in a follow-up PR.
-                return;
-            }
-            if let Ok(Some(merged)) = data::save_overrides(&url, &uuid, &overrides).await {
-                book_state.set(merged);
-            }
-        });
-    }
-}
-
-/// Build the authors-cell save callback. Mirrors [`build_save_field`] but
-/// posts a `creators` override (full replacement list) instead of a scalar.
-fn build_save_authors(
-    uuid: String,
-    server_url: String,
-    mut book_state: Signal<EbookMetadata>,
-) -> impl FnMut(Vec<String>) + 'static {
-    move |new_names: Vec<String>| {
-        let uuid = uuid.clone();
-        let url = server_url.clone();
-        spawn(async move {
-            let creators: Vec<Contributor> = new_names
-                .iter()
-                .map(|name| Contributor {
-                    name: name.clone(),
-                    ..Default::default()
-                })
-                .collect();
-            let overrides = MetadataOverrides {
-                creators: Some(creators),
-                ..Default::default()
-            };
-            if overrides.validate().is_err() {
-                return;
-            }
-            if let Ok(Some(merged)) = data::save_overrides(&url, &uuid, &overrides).await {
-                book_state.set(merged);
-            }
-        });
-    }
-}
-
-/// Inner row markup — the `<tr>` + every `<td>`. Split out of [`EbookRow`]
-/// so the parent stays a thin state-setup shell. All inputs are already
-/// derived; this component does no signal seeding of its own.
-#[component]
-fn EbookRowMarkup(
-    display: RowDisplay,
-    uuid: String,
-    is_admin: bool,
-    editing: Signal<Option<EditField>>,
-    authors_draft: Signal<Vec<String>>,
-    author_suggestions: ReadSignal<Vec<SuggestionItem>>,
-    save_field: EventHandler<(EditField, String)>,
-    save_authors: EventHandler<Vec<String>>,
-) -> Element {
-    let nav = use_navigator();
-    let uuid_click = uuid.clone();
-    let uuid_key = uuid;
-    let row_testid = display.row_testid.clone();
-    let aria_title = display.title.clone();
-
-    rsx! {
-        tr {
-            class: "ebook-row",
-            "data-testid": "{row_testid}",
-            id: "{row_testid}",
-            role: "button",
-            tabindex: "0",
-            aria_label: "Open details for {aria_title}",
-            onclick: move |_| {
-                // Row-level click navigates only when no cell-level edit
-                // is in progress. Each editable cell stops propagation on
-                // click already, but a click on a non-editable area
-                // (cover, formats, dates) while a cell is open would
-                // otherwise blur+save the input AND fire the navigation.
-                // Bailing on the editing-is-some path keeps the user on
-                // the page while the blur-save lands.
-                if editing().is_some() {
-                    return;
-                }
-                nav.push(Route::BookDetail { uuid: uuid_click.clone() });
-            },
-            onkeydown: move |evt: Event<KeyboardData>| {
-                if editing().is_some() {
-                    return;
-                }
-                let key = evt.key();
-                if key == Key::Enter || key == Key::Character(" ".to_string()) {
-                    evt.prevent_default();
-                    nav.push(Route::BookDetail { uuid: uuid_key.clone() });
-                }
-            },
-            EbookRowCells {
-                display,
-                is_admin,
-                editing,
-                authors_draft,
-                author_suggestions,
-                save_field,
-                save_authors,
-            }
-        }
-    }
-}
-
-/// Flat list of `<td>` cells inside an [`EbookRow`]. Split out of
-/// [`EbookRowMarkup`] so the row-level event handlers and the per-cell
-/// wiring live in separate functions.
-#[component]
-fn EbookRowCells(
-    display: RowDisplay,
-    is_admin: bool,
-    editing: Signal<Option<EditField>>,
-    authors_draft: Signal<Vec<String>>,
-    author_suggestions: ReadSignal<Vec<SuggestionItem>>,
-    save_field: EventHandler<(EditField, String)>,
-    save_authors: EventHandler<Vec<String>>,
-) -> Element {
-    let RowDisplay {
-        book,
-        row_testid: _,
-        thumb_base,
-        has_cover,
-        title,
-        series_line,
-        series_text,
-        authors_text,
-        updated,
-        added,
-        publisher,
-        published,
-        language,
-    } = display;
-    let cover_alt = title.clone();
-
-    rsx! {
-        EbookRowCoverCell { thumb_base, has_cover, alt_title: cover_alt }
-        RowTitleCell {
-            title,
-            error: book.error,
-            is_admin,
-            editing,
-            save_field,
-        }
-        AuthorsCell {
-            is_admin,
-            editing,
-            authors_draft,
-            authors_text,
-            suggestions: author_suggestions,
-            on_change: move |names: Vec<String>| save_authors.call(names),
-        }
-        RowSeriesCell { series_line, series_text, is_admin, editing, save_field }
-        RowScalarCell {
-            col_class: "ebook-col-publisher".to_string(),
-            cell_testid: "ebook-cell-publisher".to_string(),
-            field: EditField::Publisher,
-            value: publisher,
-            placeholder: "Publisher".to_string(),
-            is_admin,
-            editing,
-            save_field,
-        }
-        RowScalarCell {
-            col_class: "ebook-col-published".to_string(),
-            cell_testid: "ebook-cell-published".to_string(),
-            field: EditField::Published,
-            value: published,
-            placeholder: "YYYY-MM-DD".to_string(),
-            is_admin,
-            editing,
-            save_field,
-        }
-        EbookRowFormatsCell { formats: book.formats }
-        td { class: "ebook-col-updated", "data-testid": "ebook-cell-updated", "{updated}" }
-        td { class: "ebook-col-added", "data-testid": "ebook-cell-added", "{added}" }
-        RowScalarCell {
-            col_class: "ebook-col-language".to_string(),
-            cell_testid: "ebook-cell-language".to_string(),
-            field: EditField::Language,
-            value: language,
-            placeholder: "en".to_string(),
-            is_admin,
-            editing,
-            save_field,
-        }
-    }
-}
-
-/// Title cell wrapper — pre-wires the title field's [`EditableCell`] props
-/// and routes its `on_save` through the shared row save callback.
-#[component]
-fn RowTitleCell(
-    title: String,
-    error: Option<String>,
-    is_admin: bool,
-    editing: Signal<Option<EditField>>,
-    save_field: EventHandler<(EditField, String)>,
-) -> Element {
-    rsx! {
-        EditableCell {
-            col_class: "ebook-col-title".to_string(),
-            cell_testid: "ebook-cell-title".to_string(),
-            field: EditField::Title,
-            display_value: title,
-            is_admin,
-            editing,
-            placeholder: "Title".to_string(),
-            on_save: move |v: String| save_field.call((EditField::Title, v)),
-            error,
-        }
-    }
-}
-
-/// Series cell wrapper — display shows "Name #Index" (F1.3), edit input
-/// seeds with just the series name (the index lives on the full edit page).
-#[component]
-fn RowSeriesCell(
-    series_line: String,
-    series_text: String,
-    is_admin: bool,
-    editing: Signal<Option<EditField>>,
-    save_field: EventHandler<(EditField, String)>,
-) -> Element {
-    rsx! {
-        EditableCell {
-            col_class: "ebook-col-series".to_string(),
-            cell_testid: "ebook-cell-series".to_string(),
-            field: EditField::Series,
-            display_value: series_line,
-            edit_value: Some(series_text),
-            is_admin,
-            editing,
-            placeholder: "Series".to_string(),
-            on_save: move |v: String| save_field.call((EditField::Series, v)),
-            error: None,
-        }
-    }
-}
-
-/// Plain scalar-text cell wrapper for publisher / published / language —
-/// collapses the boilerplate of wiring `EditableCell` to the shared
-/// `save_field` callback.
-#[component]
-fn RowScalarCell(
-    col_class: String,
-    cell_testid: String,
-    field: EditField,
-    value: String,
-    placeholder: String,
-    is_admin: bool,
-    editing: Signal<Option<EditField>>,
-    save_field: EventHandler<(EditField, String)>,
-) -> Element {
-    rsx! {
-        EditableCell {
-            col_class,
-            cell_testid,
-            field,
-            display_value: value,
-            is_admin,
-            editing,
-            placeholder,
-            on_save: move |v: String| save_field.call((field, v)),
-            error: None,
-        }
-    }
-}
-
-/// Cover-thumbnail `<td>` — image with srcset when a cover exists, em-dash
-/// fallback otherwise.
-#[component]
-fn EbookRowCoverCell(thumb_base: String, has_cover: bool, alt_title: String) -> Element {
-    rsx! {
-        td { class: "ebook-col-cover", "data-testid": "ebook-cell-cover",
-            if has_cover {
-                img {
-                    class: "ebook-thumb",
-                    src: "{thumb_base}/md",
-                    srcset: "{thumb_base}/sm 160w, {thumb_base}/md 320w, {thumb_base}/lg 640w",
-                    sizes: "(max-width: 640px) 160px, (max-width: 1280px) 320px, 640px",
-                    alt: "Cover of {alt_title}",
-                    loading: "lazy",
-                    width: "320",
-                    height: "480",
-                }
-            } else {
-                div { class: "ebook-thumb ebook-thumb-fallback", "—" }
-            }
-        }
-    }
-}
-
-/// Formats `<td>` — one badge span per format, em-dash when empty.
-#[component]
-fn EbookRowFormatsCell(formats: Vec<String>) -> Element {
-    rsx! {
-        td { class: "ebook-col-formats", "data-testid": "ebook-cell-formats",
-            if formats.is_empty() {
-                span { class: "ebook-cell-formats-empty", "—" }
-            } else {
-                for fmt in formats.iter() {
-                    span { key: "{fmt}", class: "format-badge", "{format_badge_label(fmt)}" }
-                }
-            }
-        }
-    }
-}
-
-/// Inline-editable text cell used by [`EbookRow`]. Renders a span of
-/// text by default; in admin mode, a click swaps to a text input that
-/// commits via [`EditField::on_save`] on Enter or blur and cancels on
-/// Escape. The input's `onclick` stops propagation so the row-level
-/// navigate handler doesn't fire while the user is editing.
-///
-/// `suppress_inline_input` is the Authors-cell escape hatch: the click
-/// still toggles `editing` (so the expansion row appears) but the
-/// in-cell input never renders.
-#[component]
-fn EditableCell(
-    col_class: String,
-    cell_testid: String,
-    field: EditField,
-    display_value: String,
-    /// Separate value used to seed the input when edit mode opens.
-    /// Some cells render a richer display string than the underlying
-    /// scalar override — Series shows "Pioneers #1" but only "Pioneers"
-    /// is the series-name override. Pass `Some(bare_value)` so the
-    /// input seeds (and the blur-comparison runs against) the editable
-    /// scalar, not the rendered text. Defaults to `display_value`.
-    #[props(default)]
-    edit_value: Option<String>,
-    is_admin: bool,
-    editing: Signal<Option<EditField>>,
-    placeholder: String,
-    on_save: EventHandler<String>,
-    error: Option<String>,
-) -> Element {
-    let mut editing = editing;
-    let mut draft = use_signal(String::new);
-    let is_editing = editing() == Some(field);
-    let active_class = if is_editing {
-        " ebook-cell-editing"
-    } else {
-        ""
-    };
-    let admin_class = if is_admin { " ebook-cell-editable" } else { "" };
-    let combined_class = format!("{col_class}{admin_class}{active_class}");
-    // Single source of truth for "what the cell holds right now" —
-    // used to seed the draft on click, restore it on Escape, and skip
-    // the save-on-blur round-trip when nothing actually changed.
-    let initial = edit_value.unwrap_or_else(|| display_value.clone());
-    let initial_for_click = initial.clone();
-    let initial_for_cancel = initial.clone();
-    let initial_for_enter = initial.clone();
-    let initial_for_blur = initial.clone();
-    let testid_input = format!("{cell_testid}-input");
-
-    rsx! {
-        td {
-            class: "{combined_class}",
-            "data-testid": "{cell_testid}",
-            onclick: move |e| {
-                if !is_admin {
-                    return;
-                }
-                e.stop_propagation();
-                if !is_editing {
-                    draft.set(initial_for_click.clone());
-                    editing.set(Some(field));
-                }
-            },
-            if is_editing {
-                input {
-                    class: "ebook-cell-edit",
-                    "data-testid": "{testid_input}",
-                    autofocus: true,
-                    value: "{draft}",
-                    placeholder: "{placeholder}",
-                    onclick: move |e| { e.stop_propagation(); },
-                    oninput: move |e| { draft.set(e.value()); },
-                    onkeydown: move |e: Event<KeyboardData>| {
-                        match e.key() {
-                            Key::Enter => {
-                                e.prevent_default();
-                                // Stop the Enter event from bubbling to
-                                // the row-level keydown that navigates
-                                // to the book detail page.
-                                e.stop_propagation();
-                                let current = draft();
-                                if current.trim() != initial_for_enter.trim() {
-                                    on_save.call(current);
-                                }
-                                editing.set(None);
-                            }
-                            Key::Escape => {
-                                e.prevent_default();
-                                e.stop_propagation();
-                                draft.set(initial_for_cancel.clone());
-                                editing.set(None);
-                            }
-                            _ => {}
-                        }
-                    },
-                    onblur: move |_| {
-                        // Save on blur only when the draft actually
-                        // changed. Clicking into a cell and clicking
-                        // back out without typing must not POST an
-                        // override equal to the scanned value (which
-                        // would leak `metadata_overrides` rows that
-                        // match the underlying scan, defeating the
-                        // F5.1 merge semantics).
-                        let current = draft();
-                        if current.trim() != initial_for_blur.trim() {
-                            on_save.call(current);
-                        }
-                        editing.set(None);
-                    },
-                }
-            } else {
-                div { class: "ebook-title-cell", "{display_value}" }
-                if let Some(err) = error.as_ref() {
-                    div { class: "error", "⚠ {err}" }
-                }
-            }
-        }
-    }
-}
-
-/// Inline-editable Authors cell. Unlike [`EditableCell`] (which hosts
-/// a single-line text input), the Authors cell renders the full
-/// [`ChipEditor`] *inside* the `<td>` when the row's editing signal
-/// matches `EditField::Authors`. The cell grows vertically to fit the
-/// chips + input + dropdown, matching the design comp.
-///
-/// Display mode: comma-joined names. Hover: dashed amber outline.
-/// Click: swaps to chip editor (auto-focused) with the library-wide
-/// author pool surfaced in the dropdown. Escape exits edit mode.
-#[component]
-fn AuthorsCell(
-    is_admin: bool,
-    editing: Signal<Option<EditField>>,
-    authors_draft: Signal<Vec<String>>,
-    authors_text: String,
-    suggestions: ReadSignal<Vec<SuggestionItem>>,
-    on_change: EventHandler<Vec<String>>,
-) -> Element {
-    let mut editing = editing;
-    let is_editing = editing() == Some(EditField::Authors);
-    let active_class = if is_editing {
-        " ebook-cell-editing"
-    } else {
-        ""
-    };
-    let admin_class = if is_admin { " ebook-cell-editable" } else { "" };
-    let combined_class = format!("ebook-col-author{admin_class}{active_class}");
-
-    rsx! {
-        td {
-            class: "{combined_class}",
-            "data-testid": "ebook-cell-author",
-            onclick: move |e| {
-                if !is_admin {
-                    return;
-                }
-                e.stop_propagation();
-                if !is_editing {
-                    editing.set(Some(EditField::Authors));
-                }
-            },
-            if is_editing {
-                div { class: "ebook-cell-chip-host",
-                    ChipEditor {
-                        values: authors_draft,
-                        placeholder: "+ add author\u{2026}".to_string(),
-                        on_change,
-                        suggestions,
-                        show_avatar: false,
-                        aria_remove_prefix: "Remove".to_string(),
-                        testid_prefix: "ebook-cell-author".to_string(),
-                        autofocus: true,
-                        dropdown_header: "ADD AUTHOR".to_string(),
-                        on_close: move |_| {
-                            editing.set(None);
-                        },
-                    }
-                }
-            } else {
-                div { class: "ebook-title-cell", "{authors_text}" }
-            }
-        }
-    }
-}
